✨ 충돌 이벤트 개발
미사일<>적 상호이벤트 처리

+++ lib/src/model/missile_model.dart
@@ -5,7 +5,7 @@ class MissileModel extends Equatable {
   final double y;
   final double angle;
   final double speed;
-  final int power;
+  final double power;
 
   const MissileModel({
     this.x = 0.0,
@@ -20,7 +20,7 @@ class MissileModel extends Equatable {
     double? y,
     double? angle,
     double? speed,
-    int? power,
+    double? power,
   }) {
     return MissileModel(
       x: (x ?? this.x),
